added action methods to team controller
