+++ dropShippingApp/Controllers/TeamController.cs
@@ -22,56 +22,15 @@ namespace dropShippingApp.Controllers
 
         public async Task<IActionResult> BrowseTeams()
         {
-
+            // TODO
+            // returns team results page 
+            return View();
         }
 
         public async Task<IActionResult> SearchTeams(string searchTerm)
         {
             // TODO
-            // returns team results page
-            return View();
-        }
-
-        public async Task<IActionResult> ManageTeamProducts()
-        {
-            // TODO
-            // returns team product management page
-            return View();
-        }
-
-        public async Task<IActionResult> AddTeamProduct()
-        {
-            // TODO
-            // redirects to team product management page
-            return View();
-        }
-
-        public async Task<IActionResult> UpdateTeamProduct()
-        {
-            // TODO
-            // redirects to team product management page
-            return View();
-        }
-
-        public async Task<IActionResult> RemoveTeamProduct()
-        {
-            // TODO
-            // redirects to team product management page
-            return View();
-        }
-
-        public async Task<IActionResult> UpdateTeamSettings()
-        {
-            // TODO: will take in settings view model
-            // redirect to home management page
-            return View();
-        }
-
-        public async Task<IActionResult> UploadNewBanner()
-        {
-            // TODO: will take in formdata with an image
-            // shove image into AWS file system
-            // return home management page
+            // returns team results page (will have view model with search term)
             return View();
         }
     }
